Add `to_string_raw` argument on `MarkupField`
Sometimes it may be preferable to return the raw markup as a default when
using the `MarkupField` with something like `Django-REST-Framework`.
Previously, the field would default to the rendered value, and attempting
to source a field over to the `.raw` would require writing methods for
modifying the value held in the field.